Anget # 08

+++ src/AppAdminBundle/Controller/CRUDController.php
@@ -70,6 +70,15 @@ class CRUDController extends BaseController
         );
     }
 
+    public function batchActionDeleteIsRelevant(array $selectedIds)
+    {
+        if (!$this->getDoctrine()->getRepository('AppBundle:ProductModelSpecificSize')
+            ->isProductModelsIsOrdered($selectedIds)){
+            return true;
+        }
+        return $this->admin->trans( 'flash_delete_not_ordered_error1', [], 'AppAdminBundle' );
+    }
+
     /**
      * Batch action activate
      *
@@ -197,13 +206,19 @@ class CRUDController extends BaseController
             }elseif ($object instanceof Products) {
                 $this->addFlash(
                     'sonata_flash_error',
-                    $this->admin->trans( 'flash_delete_not_empty_error1', [ ], 'AppAdminBundle' )
+                    $this->admin->trans( 'flash_delete_not_empty_error1', [], 'AppAdminBundle' )
                 );
             }
 
             return $this->redirectTo($object);
         }
-
+         if ($this->get('product')->checkProductsIsOrdered($object)){
+             $this->addFlash(
+                 'warning',
+                 $this->admin->trans( 'flash_delete_not_ordered_error1', [], 'AppAdminBundle' )
+             );
+             return $this->redirectTo($object);
+         }
         return parent::deleteAction($id);
     }
 }
